+++ packages/affiliate/src/index.ts
@@ -1,9 +1,6 @@
-import "./install";
-
-// export * as Index from "./pages/home/index";
-// export * as Affiliate from "./pages/affiliate";
-// export * as Trader from "./pages/trader";
 export * as Dashboard from "./pages/dashboard";
+
+export * from "./pages/multiLevel";
 
 export * from "./provider";