+++ packages/affiliate/src/pages/affiliate/referralCodes/referralCodes.ui.tsx
@@ -15,8 +15,8 @@ import {
 } from "@orderly.network/ui";
 import { Decimal } from "@orderly.network/utils";
 import { EditCode } from "../../../components/editCodeBtn";
-import { EditIcon } from "../../../components/editIcon";
 import { PinBtn } from "../../../components/pinButton";
+import { EditIcon } from "../../../icons/editIcon";
 import { ReferralCodesReturns, ReferralCodeType } from "./referralCodes.script";
 
 export const ReferralCodes: FC<ReferralCodesReturns> = (props) => {