+++ packages/ui/src/index.ts
@@ -85,6 +85,7 @@ export * from "embla-carousel";
 
 export * from "./hooks";
 
+export * from "./tips";
 export * from "./locale";
 export * from "./scrollIndicator";
 export * from "./status";